Move ApplicationVersion to common package
