+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/left/LeftNavController.java
@@ -17,7 +17,6 @@
 
 package bisq.desktop.main.left;
 
-import bisq.application.ApplicationVersion;
 import bisq.bisq_easy.BisqEasyNotificationsService;
 import bisq.bisq_easy.ChatChannelDomainNavigationTargetMapper;
 import bisq.bisq_easy.NavigationTarget;
@@ -25,6 +24,7 @@ import bisq.bonded_roles.bonded_role.AuthorizedBondedRolesService;
 import bisq.chat.ChatChannelDomain;
 import bisq.chat.notifications.ChatNotification;
 import bisq.chat.notifications.ChatNotificationService;
+import bisq.common.application.ApplicationVersion;
 import bisq.common.observable.Pin;
 import bisq.desktop.ServiceProvider;
 import bisq.desktop.common.threading.UIThread;
