+++ apps/desktop/desktop/src/main/java/bisq/desktop/splash/SplashController.java
@@ -17,7 +17,7 @@
 
 package bisq.desktop.splash;
 
-import bisq.application.ApplicationVersion;
+import bisq.common.application.ApplicationVersion;
 import bisq.common.observable.Observable;
 import bisq.common.observable.Pin;
 import bisq.desktop.ServiceProvider;
