+++ common/src/main/resources/version.txt
